requirements_webapp.txt
pandas
numpy
scikit-learn
flask
flask-cors
flask-sqlalchemy
